% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waddingtonPlot.R
\name{waddingtonPlot}
\alias{waddingtonPlot}
\title{Plot Waddington Epigenetic Landscape using ggridges}
\usage{
waddingtonPlot(branches = c(1, 2, 4), horizontal_distribution = NULL,
  vertical_distribution = NULL, bottom_width = 1000, top_width = 0.3
  * bottom_width, ridge.height = 5, ridge.colors = c("red", "orange",
  "yellow", "green", "blue", "cyan", "purple"), line.color = "black",
  line.type = 1, line.size = 1, line.alpha = 0.2, theme.void = T,
  hide.legend = T, do.return = F)
}
\arguments{
\item{branches}{a vector of integer numbers of steady states at each level of the hirarchy tree or developmental path, ordered from top to bottom.}

\item{horizontal_distribution}{NULL or a list, whose length must be the same with vector `branches`, defining proportions in width of the states on the same level. (default: NULL)}

\item{vertical_distribution}{NULL or a vector, whose length must be 1 less than vector `branches`, defining vertical stretch between levels. (default: NULL)}

\item{bottom_width}{width of bottom of waddington plot}

\item{top_width}{width of top of waddington plot}

\item{ridge.height}{height of ridge}

\item{ridge.colors}{colors of redges}

\item{line.color}{color of line}

\item{line.type}{type of line}

\item{line.size}{size of line}

\item{line.alpha}{alpha of line}

\item{theme.void}{logical, whether to use the ggplot2 void theme}

\item{hide.legend}{logical, whether to hide legend}

\item{do.return}{logical, whether to return a ggplot object}
}
\value{
NULL if do.return == TRUE. otherwise a ggplot object
}
\description{
Plot Waddington Epigenetic Landscape using ggridges
}
\examples{
waddingtonPlot(branches = c(1,2,3,4),
               horizontal_distribution = list(c(1), c(1,1), c(0.5,0.5,1), c(0.25,0.25,0.5,1)))
}
